Une contrat automatique s'ajoute aprés l'insertion du joueur + Transfert bien organisée

+++ FinancialEngine.php
@@ -1,7 +1,7 @@
 <?php
 final class FinancialEngine {
     private float $commision ;
-    private int $Solvabilite ;
+    private float $Solvabilite ;
     static private float $transfertTaxe = 0.05 ;
     private float $total ;
     public function  checkSolvabilité($budget,$transfertMontant){
